requirements.txt
spotipy==2.4.4
deezer-python==1.3.0
youtube-data-api==0.0.17
soundcloud-lib==0.5.0
Cython==0.29.13
pandas==0.25.1
mongoengine==0.18.2
python_telegram_bot==12.2.0
redis==3.3.11
pyTelegramBotAPI==3.6.6
telegram==0.0.1
dnspython==1.16.0
Pillow==6.2.0
